app/requirements.txt
django>=3.2, <3.3
djangorestframework>= 3.12.4, <4.0.0
drf-spectacular>=0.15.1, <2.0.0

